# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Teste Diego Ananias Adoorei",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8021",
            "description": "Localhost api Adoorei"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of products",
                "description": "Returns list of products",
                "operationId": "getProductsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Add a new product",
                "description": "Adds a new product to the system",
                "operationId": "addProduct",
                "requestBody": {
                    "description": "Product object that needs to be added to the store",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Product X"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A new product description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product added successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Product": {
                "title": "Product",
                "description": "Product model",
                "properties": {
                    "id": {
                        "description": "ID of the product",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the product",
                        "type": "string",
                        "example": "A nice product"
                    },
                    "price": {
                        "description": "Name of the product",
                        "type": "decimal",
                        "example": "A nice product"
                    },
                    "description": {
                        "description": "Description of the product",
                        "type": "string",
                        "example": "A nice product"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Product"
                }
            }
        }
    }
}